fix: improve invalid client key error handling
