+++ ios/Components/BaseModule.swift
@@ -6,6 +6,7 @@
 
 import Adyen
 import Adyen3DS2
+import AdyenNetworking
 import React
 import UIKit
 
@@ -41,7 +42,9 @@ internal class BaseModule: RCTEventEmitter {
     internal var actionHandler: AdyenActionComponent?
 
     internal func present(_ component: PresentableComponent) {
-        guard let presenter = BaseModule.currentPresenter ?? UIViewController.topPresenter else { return sendEvent(error: NativeModuleError.notKeyWindow) }
+        guard let presenter = BaseModule.currentPresenter ?? UIViewController.topPresenter else {
+            return sendEvent(error: NativeModuleError.notKeyWindow)
+        }
 
         defer {
             BaseModule.currentPresenter = presenter
@@ -72,15 +75,8 @@ internal class BaseModule: RCTEventEmitter {
         sendEvent(withName: event.rawValue, body: [:])
     }
 
-    internal func checkErrorType(_ error: Error) -> Error {
-        if error.isComponentCanceled || error.is3DSCanceled {
-            return NativeModuleError.canceled
-        }
-        return error
-    }
-
     internal func sendEvent(error: Error) {
-        let errorToSend = checkErrorType(error)
+        let errorToSend = NativeModuleError.checkErrorType(error)
         sendEvent(withName: Events.didFail.rawValue, body: errorToSend.jsonObject)
     }
 
@@ -161,87 +157,6 @@ internal class BaseModule: RCTEventEmitter {
 
             self.currentComponent?.finalizeIfNeeded(with: result) {
                 self.cleanUp()
-            }
-        }
-    }
-}
-
-extension Error {
-
-    var isComponentCanceled: Bool { (self as? ComponentError) == ComponentError.cancelled }
-
-    var is3DSCanceled: Bool {
-        (self as NSError).domain == "com.adyen.Adyen3DS2.ADYRuntimeError" &&
-            (self as NSError).code == ADYRuntimeErrorCode.challengeCancelled.rawValue
-    }
-}
-
-extension BaseModule {
-
-    enum NativeModuleError: LocalizedError, KnownError {
-        case canceled
-        case noClientKey
-        case noPayment
-        case notSupported
-        case invalidPaymentMethods
-        case invalidAction
-        case notKeyWindow
-        case paymentMethodNotFound(PaymentMethod.Type)
-        case balanceCheck(message: String)
-        case orderRequest(message: String)
-        case sessionError
-
-        var errorCode: String {
-            switch self {
-            case .canceled:
-                return "canceledByShopper"
-            case .notSupported:
-                return "notSupported"
-            case .noClientKey:
-                return "noClientKey"
-            case .noPayment:
-                return "noPayment"
-            case .invalidPaymentMethods:
-                return "invalidPaymentMethods"
-            case .invalidAction:
-                return "invalidAction"
-            case .paymentMethodNotFound:
-                return "noPaymentMethod"
-            case .notKeyWindow:
-                return "notKeyWindow"
-            case .balanceCheck:
-                return "balanceCheck"
-            case .orderRequest:
-                return "orderRequest"
-            case .sessionError:
-                return "sessionError"
-            }
-        }
-
-        var errorDescription: String? {
-            switch self {
-            case .canceled:
-                return "Payment canceled by shopper"
-            case .notSupported:
-                return "Not supported on iOS"
-            case .noClientKey:
-                return "No clientKey in configuration"
-            case .noPayment:
-                return "No payment in configuration"
-            case .invalidPaymentMethods:
-                return "Can not parse paymentMethods or the list is empty"
-            case .invalidAction:
-                return "Can not parse action"
-            case let .paymentMethodNotFound(type):
-                return "Can not find payment method of type \(type) in provided list"
-            case .notKeyWindow:
-                return "Can not find root ViewController"
-            case let .balanceCheck(message):
-                return "Balance check error: \(message)"
-            case let .orderRequest(message):
-                return "Order request error: \(message)"
-            case .sessionError:
-                return "Something went wrong while starting session"
             }
         }
     }
