+++ ios/Components/DropInModule.swift
@@ -31,7 +31,9 @@ internal final class DropInModule: BaseModule {
     func update(_ results: NSArray) {
         guard let lookupHandler else { return }
 
-        let addressModels: [LookupAddressModel] = results.compactMap { $0 as? NSDictionary }.compactMap { try? $0.decode() }
+        let addressModels: [LookupAddressModel] = results
+            .compactMap { $0 as? NSDictionary }
+            .compactMap { try? $0.decode() }
         DispatchQueue.main.async {
             lookupHandler(addressModels)
         }
@@ -136,7 +138,9 @@ internal final class DropInModule: BaseModule {
 }
 
 extension DropInModule: DropInComponentDelegate {
-    func didSubmit(_ data: Adyen.PaymentComponentData, from component: Adyen.PaymentComponent, in dropInComponent: Adyen.AnyDropInComponent) {
+    func didSubmit(_ data: Adyen.PaymentComponentData,
+                   from component: Adyen.PaymentComponent,
+                   in dropInComponent: Adyen.AnyDropInComponent) {
         let response: SubmitData
         if let appleData = data.paymentMethod as? ApplePayDetails {
             response = SubmitData(paymentData: data.jsonObject, extra: appleData.extraData)
@@ -146,24 +150,32 @@ extension DropInModule: DropInComponentDelegate {
         sendEvent(event: .didSubmit, body: response.jsonObject)
     }
 
-    func didFail(with error: Error, from component: Adyen.PaymentComponent, in dropInComponent: Adyen.AnyDropInComponent) {
+    func didFail(with error: Error,
+                 from component: Adyen.PaymentComponent,
+                 in dropInComponent: Adyen.AnyDropInComponent) {
         sendEvent(error: error)
     }
 
-    func didProvide(_ data: Adyen.ActionComponentData, from component: Adyen.ActionComponent, in dropInComponent: Adyen.AnyDropInComponent) {
+    func didProvide(_ data: Adyen.ActionComponentData,
+                    from component: Adyen.ActionComponent,
+                    in dropInComponent: Adyen.AnyDropInComponent) {
         sendEvent(event: .didProvide, body: data.jsonObject)
     }
 
-    func didComplete(from component: Adyen.ActionComponent, in dropInComponent: Adyen.AnyDropInComponent) {
+    func didComplete(from component: Adyen.ActionComponent,
+                     in dropInComponent: Adyen.AnyDropInComponent) {
         let result = ResultDTO(result: .presentToShopper)
         sendEvent(event: .didComplete, body: result.jsonObject)
     }
 
-    func didFail(with error: Error, from component: Adyen.ActionComponent, in dropInComponent: Adyen.AnyDropInComponent) {
+    func didFail(with error: Error,
+                 from component: Adyen.ActionComponent,
+                 in dropInComponent: Adyen.AnyDropInComponent) {
         sendEvent(error: error)
     }
 
-    func didFail(with error: Error, from dropInComponent: Adyen.AnyDropInComponent) {
+    func didFail(with error: Error,
+                 from dropInComponent: Adyen.AnyDropInComponent) {
         sendEvent(error: error)
     }
 }
@@ -175,7 +187,8 @@ extension DropInModule: AddressLookupProvider {
         sendEvent(event: .didUpdateAddress, body: searchTerm)
     }
 
-    func complete(incompleteAddress: LookupAddressModel, resultHandler: @escaping (Result<PostalAddress, any Error>) -> Void) {
+    func complete(incompleteAddress: LookupAddressModel,
+                  resultHandler: @escaping (Result<PostalAddress, any Error>) -> Void) {
         lookupCompliationHandler = resultHandler
         sendEvent(event: .didConfirmAddress, body: incompleteAddress.jsonObject)
     }
@@ -183,7 +196,8 @@ extension DropInModule: AddressLookupProvider {
 }
 
 extension DropInModule: StoredPaymentMethodsDelegate {
-    func disable(storedPaymentMethod: any Adyen.StoredPaymentMethod, completion: @escaping Adyen.Completion<Bool>) {
+    func disable(storedPaymentMethod: any Adyen.StoredPaymentMethod,
+                 completion: @escaping Adyen.Completion<Bool>) {
         disableStoredPaymentMethodHandler = completion
         sendEvent(event: .didDisableStoredPaymentMethod, body: storedPaymentMethod.jsonObject)
     }
@@ -204,7 +218,9 @@ struct AddressError: Error, LocalizedError, Codable {
 
 extension DropInModule: PartialPaymentDelegate {
 
-    func checkBalance(with data: PaymentComponentData, component: any Adyen.Component, completion: @escaping (Result<Balance, any Error>) -> Void) {
+    func checkBalance(with data: PaymentComponentData,
+                      component: any Adyen.Component,
+                      completion: @escaping (Result<Balance, any Error>) -> Void) {
         sendEvent(event: .didCheckBalance, body: data.jsonObject)
         checkBalanceHandler = completion
     }
@@ -222,7 +238,8 @@ extension DropInModule: PartialPaymentDelegate {
         }
     }
 
-    func requestOrder(for component: any Adyen.Component, completion: @escaping (Result<PartialPaymentOrder, any Error>) -> Void) {
+    func requestOrder(for component: any Adyen.Component,
+                      completion: @escaping (Result<PartialPaymentOrder, any Error>) -> Void) {
         sendEvent(event: .didRequestOrder)
         requestOrderHandler = completion
     }
@@ -270,11 +287,11 @@ extension DropInModule: CardComponentDelegate {
     func didSubmit(lastFour: String, finalBIN: String, component: Adyen.CardComponent) {
         /* No Callback implemented */
     }
-    
+
     func didChangeBIN(_ value: String, component: Adyen.CardComponent) {
         sendEvent(event: .didChangeBinValue, body: value)
     }
-    
+
     func didChangeCardBrand(_ value: [Adyen.CardBrand]?, component: Adyen.CardComponent) {
         guard let value, !value.isEmpty else { return }
         let jsonData = value.map { BinLookupDataDTO(brand: $0.type.rawValue).jsonObject }
